исправил отображение главной страницы при отсутствии предложений в городе
