+++ src/domain/offer/components/city-offers/index.tsx
@@ -30,19 +30,6 @@ export default function CityOffers({
     [comparator, offers]
   );
 
-  if (sortedOffers.length === 0) {
-    return (
-      <section className="cities__no-places">
-        <div className="cities__status-wrapper tabs__content">
-          <b className="cities__status">No places to stay available</b>
-          <p className="cities__status-description">
-            We could not find any property available at the moment in Dusseldorf
-          </p>
-        </div>
-      </section>
-    );
-  }
-
   return (
     <section className="cities__places places">
       <h2 className="visually-hidden">Places</h2>
