+++ src/pages/error/index.test.tsx
@@ -1,0 +1,21 @@
+import { render, screen } from '@testing-library/react';
+import { Error } from '.';
+import { getEmptyState } from '../../config/redux/slice/error/state';
+import { getMockStoreCreator } from '../../config/redux/utils/test';
+import ROUTES from '../../domain/router/constants/ROUTES';
+import { Provider } from 'react-redux';
+import { MockAppRouter } from '../../domain/router/utils/test/components';
+
+describe(Error.name, () => {
+  const mockStoreCreator = getMockStoreCreator();
+
+  test(`should render error page on ${ROUTES.error}`, () => {
+    const store = mockStoreCreator({ error: getEmptyState() });
+    render(
+      <Provider store={store}>
+        <MockAppRouter initialEntries={[ROUTES.error]} />
+      </Provider>
+    );
+    expect(screen.getByText('App error :(')).toBeInTheDocument();
+  });
+});
